+++ lib/src/controls/better_player_cupertino_controls.dart
@@ -189,8 +189,6 @@ class _BetterPlayerCupertinoControlsState
     );
   }
 
-
-
   @override
   void dispose() {
     _dispose();