+++ lib/src/controls/better_player_material_controls.dart
@@ -173,8 +173,6 @@ class _BetterPlayerMaterialControlsState
     );
   }
 
-
-
   @override
   void dispose() {
     _dispose();