+++ lib/src/core/better_player_utils.dart
@@ -58,8 +58,7 @@ class BetterPlayerUtils {
 
   static void log(String logMessage) {
     if (!kReleaseMode) {
-      // ignore: avoid_print
-      print(logMessage);
+      debugPrint('[BetterPlayer] $logMessage');
     }
   }
 }